+++ src/views/Users/UserCreateView/UserCreateForm.js
@@ -15,8 +15,8 @@ import {
   MenuItem,
 } from '@material-ui/core';
 import { Redirect } from 'react-router';
-import Request from '../../../../requests/api/user';
-import { SYSTEM_ADMIN } from '../../../../utils/const/user_types';
+import Request from '../../../requests/api/user';
+import { SYSTEM_ADMIN } from '../../../utils/const/user_types';
 
 const useStyles = makeStyles(() => ({
   root: {},
